B07 - 비동기 로깅 적용 및 로깅 가독성있게 개선하기 (#550)
* refactor: Json Pretty Print 설정을 local, test에서만 적용하도록 변경

* feat: query pattern 변경

* refactor: file과 console 패턴 분리

* feat: AsyncAppender 도입

* refactor: utf8 제거

* refactor: AsyncAppender 설정 변경

* refactor: Config 파일 분리
