--- backend/src/main/java/botobo/core/config/ApplicationConfig.java
@@ -1,23 +1,0 @@
-package botobo.core.config;
-
-import botobo.core.ui.search.SearchArgumentResolver;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@Configuration
-public class ApplicationConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(searchArgumentResolver());
-    }
-
-    @Bean
-    public SearchArgumentResolver searchArgumentResolver() {
-        return new SearchArgumentResolver();
-    }
-}
